[ty] Simplify recursive TypeOf plumbing
