+++ crates/ty_python_semantic/src/types/cyclic.rs
@@ -26,12 +26,10 @@ use std::hash::Hash;
 use std::marker::PhantomData;
 
 use rustc_hash::{FxHashMap, FxHashSet};
-use salsa::plumbing::AsId;
 
 use crate::Db;
 use crate::FxIndexSet;
 use crate::types::Type;
-use crate::types::function::FunctionLiteral;
 
 pub(crate) type TypeTransformer<'db, Tag> = CycleDetector<Tag, Type<'db>, Type<'db>>;
 
@@ -47,40 +45,6 @@ impl<Tag> Default for TypeTransformer<'_, Tag> {
 }
 
 pub(crate) type PairVisitor<'db, Tag, C> = CycleDetector<Tag, (Type<'db>, Type<'db>), C>;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-struct RecursiveTypeNormalizationKey {
-    function_literal: salsa::Id,
-    nested: bool,
-}
-
-std::thread_local! {
-    static ACTIVE_RECURSIVE_TYPE_NORMALIZATIONS: ActiveRecursionDetector<RecursiveTypeNormalizationKey> =
-        ActiveRecursionDetector::default();
-}
-
-/// Runs recursive type normalization under a scoped guard keyed by function literal identity.
-///
-/// `TypeOf` can make a function signature refer back to the same function through many different
-/// type components. Keeping this guard scoped here lets those components keep their ordinary
-/// `recursive_type_normalized_impl(db, div, nested)` signatures.
-pub(crate) fn visit_recursive_type_normalization<R>(
-    function_literal: FunctionLiteral<'_>,
-    nested: bool,
-    on_cycle: impl FnOnce() -> R,
-    func: impl FnOnce() -> R,
-) -> R {
-    ACTIVE_RECURSIVE_TYPE_NORMALIZATIONS.with(|detector| {
-        detector.visit(
-            &RecursiveTypeNormalizationKey {
-                function_literal: function_literal.last_definition.as_id(),
-                nested,
-            },
-            on_cycle,
-            func,
-        )
-    })
-}
 
 #[derive(Debug)]
 pub struct CycleDetector<Tag, T, R> {
